refactor: Update import paths from badfish to src.badfish for consistency
