+++ tests/test_base.py
@@ -5,7 +5,8 @@ import pytest
 from aiohttp import web
 from aiohttp.test_utils import AioHTTPTestCase
 
-from src.badfish.main import BadfishException, main
+from src.badfish.main import main
+from src.badfish.helpers.exceptions import BadfishException
 from tests import config
